feat/game: Rough draft of game/user interface
